Use conda to install dependencies in Makefile
Instead of using pip install -r, use conda env update. This way we can
use the environment.yml file to reproduce the environment which
indicates which set of packages should be installed with pip and which
should be installed with conda.

Version complete data folder structure. Simplify Makefile. Update
environment.yml

Makefile reproducibility was tested on a brand new environment.