Create MiniSom2 class and restructure
Split the som.py script into som_clustering.py for applying the SOM to
a specific set of embedding vectors and som_tuning.py for applying
hyperparameter tuning for a SOM with a specific set of embedding
vectors.
Make use of the MiniSom2 class which wraps the MiniSom class and
provides an api more similar to sklearn and additional features such
as the u_matrix method. In the future we can use this wrapper as a
base for implementing additional features such as the multiprocessing
training.