Add notebook to experiment with bokeh U-Matrix
This notebook contains two bokeh U-Matrix visualization. One composed
of single hexagons and another made of a hexagonal tile. In the future,
the former will be excluded as the later offers better capabilities.

Also the U-Matrix features an overaly of marker symbols representing
the presence of each document category in each cell in the grid.