Create Mapintel Haystack Docker Compose app
Adapt the Haystack Docker Compose application in https://github.com/deepset-ai/haystack
for the purposes of the Mapintel project.
The docker-compose.yml file defines the multi-container application. It
is composed of a container for the document store, a container for the
api built in Python with FastAPI, Haystack and GPU support, and a
container for the ui built in Python with Streamlit.
The api folder contains the code for the api. The api is defined
according to the Haystack framework (indexing pipeline, document store
and search pipeline) and consists of 5 endpoints:
    - /news-upload - used to request updated news articles from NewsAPI
    and index them in the document store;
    - /file-upload - used to index user uploaded files in the document
    store;
    - /query - used to query the document store according to the search
    pipeline;
    - /feedback - used to provide feedback on the results from the
    search pipeline;
    - /eval-feedback - used to evaluate the search pipeline based on the
    feedback provided so far.
The Haystack components we user are:
    - Document Store: Open Distro Elasticsearch;
    - Indexing Pipeline: Uses EmbeddingRetriever based on SBERT to
    create document embeddings and index them;
    - Search Pipeline: Retrieve and Re-Rank pipeline based on SBERT
    inspired by https://www.sbert.net/examples/applications/retrieve_rerank/README.html#retrieve-re-rank;
The ui folder defines the user interface. The ui is defined according to
Streamlit and integrates the api endpoints into an easy to use
application that provides users the ability to upload new files to the
document store, to query the document store, and to provide feedback to
the documents retrieved by the query (which is stored in the document
store).