Set root as docker context and reestructure code
docker-compose.yml will now build service api and ui with context set as
root. This enables each container to use the src, data and outputs
folder in their building.
The models folder was renamed to outputs as the name is more
appropriate and the code was adapted as such.
The folder api/backups/ was moved to data/backups.