Create umap and search-binary endpoints
Create the actual endpoints (umap-training, umap-inference, umap-query,
search-binary) and override/create new methods for the
OpenDistroElasticsearchDocumentStore2 custom pipeline node to
accomodate requirements of the new endpoints (
_get_all_documents_in_index, get_all_documents_generator,
get_all_documents).
The new endpoints will allow interaction with the UMAP model and
accquiring the respective embeddings for the documents in the document
store.