Replace search-binary and create doc-count apis
The new all-docs-generator endpoint replaces the old search-binary. This
new endpoint will return a StreamingResponse which consists of a
generator over the documents of the doc store. This is necessary to
prevent loading all documents in memory.
The doc-count endpoint gives the number of documents in the doc store
that satisfy a filter condition.