Adding new api endpoints and improving the UI
Add the new api endpoints (doc-count, umap-inference, umap-query,
all-docs-generator) to be used by the UI.

Integrate João's work by plotting a UMAP of the documents using the
streamlit-bokeh component to allow bidirectional interactivity.

The all-docs-generator endpoint is used to gather the data necessary to
plot the UMAP of the selected documents. umap-query is used to plot the
query embedding alongside the documents in the UMAP. doc-count is used
to obtain the number of the selected documents so we can define the
sample size given the percentage of documents to be selected. The
umap-inference endpoint is used after uploading external documents in
the doc store so we immediately obtain the corresponding umap
embeddings.