Integrate BERTopic into the system
Define the _BERTopicEncoder class to use with the TopicRetriever, which
is responsible for using BERTopic in the Retriever.

Change paths in multiple files to make them more robust to different
OSes and systems. Also, improve the reproducibility and organization by
fixing the import of custom modules.

Add search query to select random document form document store in
elasticsearch_exploration.ipynb.