Improve UMAP plot and UI
Remove bidirectional interactivity of bokeh plot as we weren't using it
anymore. This made the UI code simpler and more manageable.

The scatter plot legend is now interactive, meaning that if we select a
given label, the corresponding points will be hidden.

We can now select whether the filtered categories are to be excluded or
included.