Save title, description and content field separate
Cleaning functions in text_cleaner.py now concatenate the text of the
three fields using a separator tag that can be later used to obtain the
original fields (or replaced by a white space to get a clean
concatenated string). This is useful for the UI to display each piece
of text separately.